+++ source/minor_det.cpp
@@ -1,14 +1,14 @@
 #include "setup.h"
 
 matrix matrix::make_addition(int i, int j) {
-    double numbers[(size-1)*(size-1)];
+    long double numbers[(size-1)*(size-1)];
     int pos = 0;
 
     for (int n = 0; n < size; ++n) {
         if (n == i) continue;
         else {
-            for(int m = 0; m < size; ++m) {
-                if(m == j) continue;
+            for (int m = 0; m < size; ++m) {
+                if (m == j) continue;
                 else {
                     numbers[pos] = data[n*size + m];
                     ++pos;
@@ -20,14 +20,14 @@ matrix matrix::make_addition(int i, int j) {
     return addition;
 }
 
-double matrix::minor_det() {
+long double matrix::minor_det() {
     if (size == 1) return (data[0]);
 
     matrix addition = make_addition(0, 0); 
-    double result = 0;
+    long double result = 0;
     int sign = 1;
 
-    for(int j = 0; j < size; ++j) {
+    for (int j = 0; j < size; ++j) {
         result += (sign)*(data[j])*(addition.minor_det());
         sign *= (-1);
         if((j + 1) != size) {
